Thank me later
Lel, 500 files, Solid number. Enjoy people of RageZone! (cash shop
included)

+++ src/handling/channel/handler/MobHandler.java
@@ -62,8 +62,8 @@ public class MobHandler {
 
             if (size > 0) {
                 Pair skillToUse = (Pair) monster.getSkills().get((byte) Randomizer.nextInt(size));
-                realskill = ((Integer) skillToUse.getLeft()).intValue();
-                level = ((Integer) skillToUse.getRight()).intValue();
+                realskill = ((Integer) skillToUse.getLeft());
+                level = ((Integer) skillToUse.getRight());
 
                 MobSkill mobSkill = MobSkillFactory.getMobSkill(realskill, level);
